backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
anthropic>=0.40.0
python-multipart>=0.0.20
Pillow>=11.1.0
pydantic>=2.11.0
python-dotenv>=1.0.1
